paf/app/requirements.txt
Jinja2==3
Flask~=2.0.1
Flask-WTF==0.15.1
flask-sqlalchemy==2.5.1
flask_bootstrap==3.3.7.1
flask_fontawesome==0.1.5
flask-nav==0.6
pymysql==1.0.2
pytest==6.2.5
bradocs4py==1.3.1.0
https://github.com/TadaSoftware/PyNFe/archive/master.zip
weasyprint==54.0
Flask-QRcode==3.0.0
redis==4.1.1
Flask-Session==0.4.0
pyserial==3.5
PyJWT==1.7.1
cryptography==3.2
Authlib==0.15.2
pytz==2021.3
WTForms==2.3.3
WTForms-Ext==0.5